analyse/requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
numpy>=1.20.0
requests>=2.25.0
Pillow>=8.0.0
tqdm>=4.60.0
jupyter>=1.0.0
ipykernel>=6.0.0
flask>=2.0.0
plotly>=5.5.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
flask-caching>=2.0.0
